+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- feature: add `WrapperObject::getPropertyValue`
+
 ## 0.12.5 - 2022-09-08
 
 - fix: ignore unavailable source code when parsing property-read tags